YogaGPT/requirements.txt
discord==2.3.2
openai==1.47.1
python-dotenv==1.0.1